chore: move headers under react-app-rewired
