+++ react-app-rewired/headers/csps/assetService/trustwallet.ts
