+++ react-app-rewired/headers/csps/base.ts
