+++ react-app-rewired/headers/csps/chains/bitcoin.ts
