+++ react-app-rewired/headers/csps/defi/foxy.ts
