+++ react-app-rewired/headers/csps/marketService/coincap.ts
