+++ react-app-rewired/headers/csps/marketService/coingecko.ts
