+++ react-app-rewired/headers/csps/marketService/exchangeRates.ts
