+++ react-app-rewired/headers/csps/plugins/foxPage.ts
