+++ react-app-rewired/headers/csps/wallets/metamask.ts
